refactor: 프로젝트 디렉토리 구조 변경
모듈형 디렉토리 구조에서 계층형 디렉토리 구조로 변경했습니다. 저희 서비스가 굉장히 큰 스펙이라면 모듈형 디렉토리 구조로 변경해서 MSA에 대비하는게 좋지만 현재 그런 상황은 아니기 때문에 오히려 프로젝트 복잡도만 늘어난다고 판단하여 계층형 구조로 가는게 좋겠다고 판단했습니다.
