+++ src/main/java/service/tmbmbackend/repository/CCTVRepository.java
@@ -1,4 +1,4 @@
-package service.tmbmbackend.cctv.repository;
+package service.tmbmbackend.repository;
 
 import java.util.List;
 import org.locationtech.jts.geom.Geometry;
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
-import service.tmbmbackend.cctv.CCTV;
+import service.tmbmbackend.entity.CCTV;
 
 @Repository
 public interface CCTVRepository extends JpaRepository<CCTV, Long> {
